requirements.txt
openai>=1.12.0
python-dotenv
faiss-cpu
pypdf
tiktoken
langchain
streamlit
fastapi
uvicorn
neo4j
streamlit-agraph
pyyaml
python-multipart
dvc
rouge-score
nltk
prometheus-client
prometheus-fastapi-instrumentator